Add the VariableAnalysis snippets. Remove the abstract SchemaBasedMapper
