+++ includes/Core/Interfaces/ProductMapperInterface.php
@@ -17,7 +17,6 @@ if ( ! defined( 'ABSPATH' ) ) {
  * Product mapper interface
  */
 interface ProductMapperInterface {
-
 	/**
 	 * Map a product to feed data.
 	 *
@@ -25,5 +24,5 @@ interface ProductMapperInterface {
 	 * @param \WC_Product|null $parent_product The parent product (for variations).
 	 * @return array Mapped product data.
 	 */
-	public function map_product_by_schema( \WC_Product $product, ?\WC_Product $parent_product = null ): array;
+	public function map_product( \WC_Product $product, ?\WC_Product $parent_product ): array;
 }
